+++ tests/TestCase/Utility/ThumbCreatorSaveTest.php
@@ -12,11 +12,14 @@
  */
 namespace Thumber\Test\TestCase\Utility;
 
+use BadMethodCallException;
 use Cake\Core\Configure;
 use Intervention\Image\Exception\InvalidArgumentException;
 use Intervention\Image\Exception\NotSupportedException;
 use RuntimeException;
 use Thumber\TestSuite\TestCase;
+use Thumber\Utility\ThumbCreator;
+use Tools\Exception\NotWritableException;
 
 /**
  * ThumbCreatorSaveTest class
@@ -62,12 +65,12 @@ class ThumbCreatorSaveTest extends TestCase
         }
 
         //With an invalid file as input.
-        $expectExceptionMessage = 'Unable to read image from file `tests/test_app/config/routes.php`';
+        $expectMessage = 'Unable to read image from file `tests/test_app/config/routes.php`';
         if (Configure::readOrFail('Thumber.driver') != 'imagick') {
-            $expectExceptionMessage = 'Image type `text/x-php` is not supported by this driver';
+            $expectMessage = 'Image type `text/x-php` is not supported by this driver';
         }
         $this->expectException(RuntimeException::class);
-        $this->expectExceptionMessage($expectExceptionMessage);
+        $this->expectExceptionMessage($expectMessage);
         $this->getThumbCreatorInstanceWithSave(APP . 'config' . DS . 'routes.php');
     }
 
@@ -106,6 +109,23 @@ class ThumbCreatorSaveTest extends TestCase
         sleep(1);
         $newTime = filemtime($this->getThumbCreatorInstance()->resize(200)->save());
         $this->assertNotEquals($time, $newTime);
+    }
+
+    /**
+     * Test for `save()` method, if unable to create file
+     * @requires OS Linux
+     * @test
+     */
+    public function testSaveUnableToCreateFile()
+    {
+        $this->expectException(NotWritableException::class);
+        $this->expectExceptionMessage('Unable to create file `' . DS . 'noExisting`');
+        $ThumbCreator = $this->getMockBuilder(ThumbCreator::class)
+            ->setConstructorArgs(['400x400.jpg'])
+            ->setMethods(['getPath'])
+            ->getMock();
+        $ThumbCreator->method('getPath')->willReturn(DS . 'noExisting');
+        $ThumbCreator->resize(200)->save();
     }
 
     /**
@@ -171,8 +191,8 @@ class ThumbCreatorSaveTest extends TestCase
      */
     public function testSaveWithoutCallbacks()
     {
-        $this->expectException(RuntimeException::class);
-        $this->expectExceptionMessage('No valid method called before the `save` method');
+        $this->expectException(BadMethodCallException::class);
+        $this->expectExceptionMessage('No valid method called before the `save()` method');
         $this->getThumbCreatorInstance()->save();
     }
 }