+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # 1.x branch
 ## 1.7 branch
+### 1.7.6
+* `ThumbCreator` can throw a `BadMethodCallException` or a `NotWritableException`;
+* `ThumbHelper` throws a `BadMethodCallException` if a method doesn't exist;
+* fixed little bug for `ThumbsPathTrait::getPath()` method.
+
 ### 1.7.5
 * fixed little bug for `ThumbsPathTrait::resolveFilePath()` method;
 * added tests for lower dependencies;